Fix solveHuber! call in QuadTuningSearch to disable debug capture
- Add debug_capture = false parameter to prevent capturing debug info during quad tuning
- Maintains performance by avoiding unnecessary debug data collection in this search phase

+++ src/Routines/SearchDIA/SearchMethods/HuberTuningSearch/utils.jl
@@ -414,7 +414,7 @@ function process_delta_values!(
             params.max_iter_outer,
             params.accuracy_newton,
             params.accuracy_bisection,
-            10.0,
+            Float32(10.0),
             params.max_diff,
             params.reg_type,
         )
