+++ src/Routines/SearchDIA/SearchMethods/IntegrateChromatogramsSearch/utils.jl
@@ -331,7 +331,7 @@ function build_chromatograms(
                 params.max_iter_outer,
                 search_context.deconvolution_stop_tolerance[],#params.accuracy_newton,
                 search_context.deconvolution_stop_tolerance[],#params.accuracy_bisection,
-                search_context.deconvolution_stop_tolerance[],
+                Float32(search_context.deconvolution_stop_tolerance[]),
                 params.max_diff,
                 params.reg_type
             )
@@ -558,7 +558,7 @@ function build_chromatograms(
                 params.max_iter_outer,
                 search_context.deconvolution_stop_tolerance[],#params.accuracy_newton,
                 search_context.deconvolution_stop_tolerance[],#params.accuracy_bisection,
-                search_context.deconvolution_stop_tolerance[],
+                Float32(search_context.deconvolution_stop_tolerance[]),
                 params.max_diff,
                 params.reg_type,#NoNorm()
             )
